@@ -1,8 +1,9 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
 
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+// const lightCodeTheme = require('prism-react-renderer/themes/github');
+// const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+import { themes as prismThemes } from 'prism-react-renderer';
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
@@ -19,7 +20,14 @@ const config = {
   organizationName: 'EarthMC', // Usually your GitHub org/user name.
   projectName: 'earthmc-docs', // Usually your repo name.
 
-  scripts: [{src: 'https://plausible.io/js/plausible.js', async: false, defer: true, 'data-domain': 'earthmc.net'}],
+  scripts: [
+    {
+      src: 'https://plausible.io/js/plausible.js',
+      async: false,
+      defer: true,
+      'data-domain': 'earthmc.net',
+    },
+  ],
 
   // Even if you don't use internalization, you can use this field to set useful
   // metadata like html lang. For example, if your site is Chinese, you may want
@@ -34,9 +42,9 @@ const config = {
       '@docusaurus/theme-classic',
       /** @type {import('@docusaurus/theme-classic').Options} */
       {
-        customCss: require.resolve('./src/css/custom.css')
+        customCss: require.resolve('./src/css/custom.css'),
       },
-    ]
+    ],
   ],
 
   plugins: [
@@ -44,78 +52,77 @@ const config = {
       '@docusaurus/plugin-content-docs',
       {
         path: 'src/docs',
-        routeBasePath: '/'
-      }
+        routeBasePath: '/',
+      },
     ],
     [
       '@docusaurus/plugin-content-pages',
       {
         path: 'src/pages',
-        routeBasePath: '/pages'
-      }
+        routeBasePath: '/pages',
+      },
     ],
-     [
+    [
       '@docusaurus/plugin-content-blog',
       {
         path: 'src/posts',
-        routeBasePath: '/posts'
-      }
+        routeBasePath: '/posts',
+      },
     ],
     [
       '@docusaurus/plugin-sitemap',
       {
         changefreq: 'weekly',
         priority: 0.5,
-      }
-    ]
+      },
+    ],
   ],
 
-  themeConfig:
-    ({
-      navbar: {
-        title: 'EarthMC docs',
-        // logo: {
-        //   alt: 'EarthMC Logo',
-        //   src: 'img/logo.svg',
-        // },
-        items: [],
-      },
-      footer: {
-        style: 'dark',
-        links: [
-          {
-            title: 'IP: play.earthmc.net',
-          },
-          {
-            title: 'Links',
-            items: [
-              {
-                label: 'Discord',
-                href: 'https://discord.gg/TFVYpWQ',
-              },
-              {
-                label: 'Wiki',
-                href: 'https://earthmc.fandom.com/',
-              },
-            ],
-          },
-          {
-            title: 'More',
-            items: [
-              {
-                label: 'GitHub',
-                href: 'https://github.com/EarthMC/',
-              },
-            ],
-          },
-        ],
-        copyright: `Copyright © ${new Date().getFullYear()} EarthMC.net.`,
-      },
-      prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
-      },
-    }),
+  themeConfig: {
+    navbar: {
+      title: 'EarthMC docs',
+      // logo: {
+      //   alt: 'EarthMC Logo',
+      //   src: 'img/logo.svg',
+      // },
+      items: [],
+    },
+    footer: {
+      style: 'dark',
+      links: [
+        {
+          title: 'IP: play.earthmc.net',
+        },
+        {
+          title: 'Links',
+          items: [
+            {
+              label: 'Discord',
+              href: 'https://discord.gg/TFVYpWQ',
+            },
+            {
+              label: 'Wiki',
+              href: 'https://wiki.earthmc.net/',
+            },
+          ],
+        },
+        {
+          title: 'More',
+          items: [
+            {
+              label: 'GitHub',
+              href: 'https://github.com/EarthMC/',
+            },
+          ],
+        },
+      ],
+      copyright: `Copyright © ${new Date().getFullYear()} EarthMC.net.`,
+    },
+    prism: {
+      theme: prismThemes.github,
+      darkTheme: prismThemes.dracula,
+    },
+  },
 };
 
 module.exports = config;
